Move the class to its own file
