+++ freezeyt/filesaver.py
@@ -4,7 +4,7 @@ from pathlib import Path, PurePosixPath
 
 from . import compat
 from .saver import Saver
-from .util import AbsoluteURL
+from .absolute_url import AbsoluteURL
 
 from typing import Callable, Iterable, BinaryIO
